For Test

+++ app/Database/Migrations/20200228154900_create_vitals.php
@@ -22,15 +22,15 @@ class CreateVitals extends \CodeIgniter\Database\Migration {
           'comment' => ''
         ],
         'weight' => [
-          'type' => 'int',
+          'type' => 'float',
           'comment' => ''
         ],
         'height' => [
-          'type' => 'int',
+          'type' => 'float',
           'comment' => ''
         ],
         'temperature' => [
-          'type' => 'int',
+          'type' => 'float',
           'comment' => ''
         ],
         'respiratory_rate' => [
